Feat(post-service) : PostLikeService 수정
게시글을 삭제하면서 좋아요도 있으면 삭제하기 때문에 예외처리를 하는 게
아님

+++ post-service/src/main/java/com/sgdevcamp/postservice/service/PostLikeService.java
@@ -45,9 +45,6 @@ public class PostLikeService {
     }
 
     public void deleteAllPostLike(String post_id){
-
-        if(postLikeRepository.findByPostId(post_id).isEmpty()) throw new CustomException(NOT_FOUND_POSTLIKE);
-
-        postLikeRepository.deleteAllInBatchByPostId(post_id);
+        if(!postLikeRepository.findByPostId(post_id).isEmpty()) postLikeRepository.deleteAllInBatchByPostId(post_id);
     }
 }
